neural-net/requirements.txt
torch>=2.0
scikit-learn>=1.3
scipy>=1.11
pyarrow>=12.0
tqdm>=4.65
seaborn>=0.12
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
requests>=2.28
python-dotenv>=1.0
